Back to registering - works better
