+++ bundles/org.openhab.binding.wizlighting/src/main/java/org/openhab/binding/wizlighting/internal/discovery/WizLightingDiscoveryService.java
@@ -28,9 +28,9 @@ import org.eclipse.smarthome.config.discovery.AbstractDiscoveryService;
 import org.eclipse.smarthome.config.discovery.DiscoveryResult;
 import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
 import org.eclipse.smarthome.config.discovery.DiscoveryService;
-import org.eclipse.smarthome.core.net.NetworkAddressService;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
+import org.openhab.binding.wizlighting.internal.entities.RegistrationRequestParam;
 import org.openhab.binding.wizlighting.internal.entities.WizLightingRequest;
 import org.openhab.binding.wizlighting.internal.enums.WizLightingMethodType;
 import org.openhab.binding.wizlighting.internal.handler.WizLightingMediator;
@@ -56,9 +56,6 @@ public class WizLightingDiscoveryService extends AbstractDiscoveryService {
 
     @NonNullByDefault({})
     private WizLightingMediator mediator;
-
-    @NonNullByDefault({})
-    private NetworkAddressService networkAddressService;
 
     private final WizLightingPacketConverter converter = new WizLightingPacketConverter();
 
@@ -89,27 +86,6 @@ public class WizLightingDiscoveryService extends AbstractDiscoveryService {
         }
     }
 
-    @Reference(cardinality = ReferenceCardinality.MANDATORY, policy = ReferencePolicy.STATIC)
-    public void setNetworkAddressService(NetworkAddressService networkAddressService) {
-        this.networkAddressService = networkAddressService;
-        logger.trace("Network Address Service has been set in the mediator.");
-    }
-
-    public void unsetNetworkAddressService(NetworkAddressService networkAddressService) {
-        this.networkAddressService = null;
-        logger.trace("Network Address Service has been unset from the mediator.");
-    }
-
-    private String getMyIpAddress() {
-        String myIpAddress = networkAddressService.getPrimaryIpv4HostAddress();
-        if (myIpAddress == null) {
-            logger.warn("Network interface did not return an IP address!");
-            return "OHIPAddress";
-        }
-        logger.info("IP of openHAB device is {}.", myIpAddress);
-        return myIpAddress;
-    }
-
     /**
      * Constructor of the discovery service.
      *
@@ -128,11 +104,12 @@ public class WizLightingDiscoveryService extends AbstractDiscoveryService {
     protected void startScan() {
         DatagramSocket dsocket = null;
         try {
-            String outIpReal = getMyIpAddress();
+            RegistrationRequestParam outParam = this.mediator.getRegistrationParams();
+            String outIpReal = outParam.getPhoneIp();
             String broadcastIp = outIpReal.substring(0, outIpReal.lastIndexOf(".")) + "255";
             InetAddress address = InetAddress.getByName(broadcastIp);
-            // Simply requesting the current state from any bulb that's listening
-            WizLightingRequest request = new WizLightingRequest(WizLightingMethodType.getPilot, null);
+            WizLightingRequest request = new WizLightingRequest(WizLightingMethodType.registration,
+                    this.mediator.getRegistrationParams());
             request.setId(0);
 
             byte[] message = this.converter.transformToByteMessage(request);
