+++ bundles/org.openhab.binding.wizlighting/src/main/java/org/openhab/binding/wizlighting/internal/handler/WizLightingMediatorImpl.java
@@ -19,13 +19,19 @@ import java.util.Set;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
+import org.eclipse.smarthome.core.net.NetworkAddressService;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.openhab.binding.wizlighting.internal.WizLightingBindingConstants;
 import org.openhab.binding.wizlighting.internal.discovery.WizLightingDiscoveryService;
+import org.openhab.binding.wizlighting.internal.entities.RegistrationRequestParam;
 import org.openhab.binding.wizlighting.internal.entities.WizLightingResponse;
 import org.openhab.binding.wizlighting.internal.runnable.WizLightingUpdateReceiverRunnable;
+import org.openhab.binding.wizlighting.internal.utils.NetworkUtils;
 import org.osgi.service.component.ComponentContext;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -47,6 +53,9 @@ public class WizLightingMediatorImpl implements WizLightingMediator {
     private @Nullable Thread receiverThread;
 
     private @Nullable WizLightingDiscoveryService wizlightingDiscoveryService;
+
+    @NonNullByDefault({})
+    private NetworkAddressService networkAddressService;
 
     /**
      * Called at the service activation.
@@ -184,6 +193,52 @@ public class WizLightingMediatorImpl implements WizLightingMediator {
     @Override
     public Set<Thing> getAllThingsRegistered() {
         return this.handlersRegisteredByThing.keySet();
+    }
+
+    @Reference(cardinality = ReferenceCardinality.MANDATORY, policy = ReferencePolicy.STATIC)
+    public void setNetworkAddressService(NetworkAddressService networkAddressService) {
+        this.networkAddressService = networkAddressService;
+        logger.trace("Network Address Service has been set in the mediator.");
+    }
+
+    public void unsetNetworkAddressService(NetworkAddressService networkAddressService) {
+        this.networkAddressService = null;
+        logger.trace("Network Address Service has been unset from the mediator.");
+    }
+
+    private String getMyIpAddress() {
+        String myIpAddress = networkAddressService.getPrimaryIpv4HostAddress();
+        if (myIpAddress == null) {
+            logger.warn("Network interface did not return an IP address!");
+            return "OHIPAddress";
+        }
+        logger.info("IP of OpenHab device is {}.", myIpAddress);
+        return myIpAddress;
+    }
+
+    private String getMyMacAddress() {
+        String myMacAddress;
+        try {
+            myMacAddress = NetworkUtils.getMyMacAddress();
+            if (myMacAddress == null) {
+                logger.warn("No network interface could be found.  Mac of OpenHab device is unknown.");
+                return "OHMACAddress";
+            }
+        } catch (Exception e) {
+            logger.warn("Mac Address of OpenHab device is invalid.");
+            return "OHMACAddress";
+        }
+        logger.info("Mac Address of OpenHab device is {}.", myMacAddress);
+        return myMacAddress;
+    }
+
+    /**
+     * Returns a {@link RegistrationRequestParam} based on the current OpenHAB
+     * connection.
+     *
+     */
+    public RegistrationRequestParam getRegistrationParams() {
+        return new RegistrationRequestParam(getMyIpAddress(), true, 0, getMyMacAddress());
     }
 
     @Override
